requirements.txt
asgiref==3.7.2
coverage==7.3.2
Django==4.2.6
django-bootstrap5==23.3
django-cleanup==8.0.0
django-cors-headers==4.3.0
django-debug-toolbar==4.2.0
django-nose==1.4.7
django-ranged-response==0.2.0
django-simple-captcha==0.5.20
djangorestframework==3.14.0
easy-thumbnails==2.8.5
mysqlclient==2.2.0
nose==1.3.7
Pillow==10.1.0
psycopg2==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
